refactor: give effect/ foundry/ + logic/ folder parity
Split the single SohlActiveEffect.ts into the standard layout (no SohlLogic
class — ActiveEffect has no actions/lifecycle; its behavior is change
application via the Foundry effect pipeline).

- foundry/SohlActiveEffect.ts — the document + the Foundry-coupled change
  dispatchers (dispatchModifierChange / dispatchStrikeModeChange, which resolve
  paths via foundry.utils.getProperty).
- foundry/SohlActiveEffectDataModel.ts — the data model + schema.
- foundry/SohlActiveEffectSheet.ts — the ActiveEffectConfig-based editor.
- logic/effect-logic.ts — the pure change-mapping helpers (changeTypeToOperator,
  pushDeltaToValueModifier); added to the ESLint Foundry-free zone + purity test.
- Update importers (SohlSystem split into three; type importers repointed).

Also convert the remaining bare `Hooks.on`/`Hooks.once` calls in SohlHookBridge
and sohl.ts to `(Hooks as any)` (the convention used elsewhere) — the added
module surface tipped tsc's `Hooks.on` overload union past its complexity limit
(TS2590); the cast sidesteps the giant union.

No behavior change.
