+++ eslint.config.js
@@ -64,6 +64,7 @@ const FOUNDRY_FREE_ZONES = [
     "src/document/combat/logic/**/*.ts",
     "src/document/token/logic/**/*.ts",
     "src/document/scene/logic/**/*.ts",
+    "src/document/effect/logic/**/*.ts",
     "src/domain/**/*.ts",
     "src/core/SohlLogic.ts",
     "src/core/SohlActionContext.ts",
